+++ packages/core/src/vuePlugins/vuex.ts
@@ -73,6 +73,7 @@ const getInitialState = (): CoreState => ({
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   configuration: {},
+  hasSmallDisplay: false,
   errors: [],
 })