+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 
 - Feature: Add globally usable css class `polar-tooltip` intended to update the appearance of `v-tooltip`s to be more visible.
+- Feature: Add new state variable `hasSmallDisplay` which is updated on `resize` of the `window`.
 
 ## 1.1.0